Fixed DEPTP-5 and DEPTP-6

+++ service/src/main/java/com/wso2telco/dep/tpservice/dao/TokenDAO.java
@@ -102,7 +102,7 @@ public class TokenDAO {
 			int id = (Integer)resultsMap.get("tokendid");
 			int whoId = (Integer)resultsMap.get("tsxwhodid");
 			String tokenAuth = (String)resultsMap.get("tokenauth");
-			int tokenValidity = (Integer)resultsMap.get("tokenvalidity");
+			Long tokenValidity = (Long)resultsMap.get("tokenvalidity");
 			String accessToken = (String)resultsMap.get("accesstoken");
 			String refreshToken = (String)resultsMap.get("refreshtoken");
 			//Timestamp lastRefreshDone = (Timestamp)resultsMap.get("lastrefreshdone");
